ContantApp/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7

# Database
motor==3.3.2
pymongo[srv]==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
google-auth==2.27.0

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.1.0

# AI Integration
openai==1.12.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
aioresponses==0.7.6
mongomock==4.1.2